fix: resolve bugs

+++ src/modules/equipment/controller/equipmentController.js
@@ -1,4 +1,5 @@
 import { equipmentModel } from "../model/equipmentModel.js"
+import {v2} from "cloudinary"
 
 export const equipmentController = {
     getEquipment: async(req, res) => {
@@ -13,11 +14,15 @@ export const equipmentController = {
     },
     addEquipment: async(req, res) => {
         try {
-                const equipmentReq = req.body
+                const equipmentReq = JSON.stringify(req.body.data)
                 if(!equipmentReq) return res.status(400).json({status:400, message:"Bad request - al insertar un equipamento", data: null})
                 const equipmentInserted = new equipmentModel(equipmentReq)
                 equipmentInserted.isAvailable = true
                 equipmentInserted.status = true
+                if(req.file){
+                    equipmentInserted.image = req.file.path
+                    equipmentInserted.imagePublicId = req.file.filename
+                }
                 const equipmentSaved = await equipmentInserted.save()
                 return res.status(201).json({status:201, message:"equipamento registrada exitosamente", data: equipmentSaved})
         } catch (error) {
@@ -40,6 +45,11 @@ export const equipmentController = {
     deleteEquipment: async(req, res) => {
         try {
                 const id = req.params.id
+                const equipmentFound = await equipmentModel.findById(id)
+                if(!equipmentFound) return res.status(404).json({status:404, message:"Paciente no encontrado", data: null})
+                if(requestCode.imagePublicId){
+                     await v2.uploader.destroy(patientFound.imagePublicId)
+                }
                 const equipmentDeleted = await equipmentModel.findByIdAndDelete(id)
                 return res.status(201).json({status:201, message:"equipo eliminada exitosamente", data: equipmentDeleted})
         } catch (error) {
